Move `File` to `Files` directory
