+++ src/Files/File.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace BrianHenryIE\Strauss;
+namespace BrianHenryIE\Strauss\Files;
 
 use BrianHenryIE\Strauss\Composer\ComposerPackage;
+use BrianHenryIE\Strauss\DiscoveredSymbol;
 
 class File
 {
