+++ tests/Unit/PrefixerTest.php
@@ -11,15 +11,14 @@ namespace BrianHenryIE\Strauss\Tests\Unit;
 
 use BrianHenryIE\Strauss\Composer\Extra\StraussConfig;
 use BrianHenryIE\Strauss\DiscoveredSymbols;
-use BrianHenryIE\Strauss\File;
+use BrianHenryIE\Strauss\Files\File;
 use BrianHenryIE\Strauss\Prefixer;
+use BrianHenryIE\Strauss\TestCase;
 use BrianHenryIE\Strauss\Types\ClassSymbol;
 use BrianHenryIE\Strauss\Types\ConstantSymbol;
 use BrianHenryIE\Strauss\Types\NamespaceSymbol;
 use Composer\Composer;
 use Composer\Config;
-use BrianHenryIE\Strauss\TestCase;
-use Mockery\Mock;
 use Symfony\Component\Console\Input\InputInterface;
 
 /**
